requirements.txt
fastmcp
httpx
openai
python-dotenv
PyYAML
